Split metal runtime and codegen
