+++ taichi/codegen/metal/env_config.cpp
@@ -1,4 +1,4 @@
-#include "taichi/backends/metal/env_config.h"
+#include "taichi/codegen/metal/env_config.h"
 
 #include "taichi/lang_util.h"
 #include "taichi/util/environ_config.h"
