+++ taichi/codegen/metal/shaders/ad_stack.metal.h
@@ -1,4 +1,4 @@
-#include "taichi/backends/metal/shaders/prolog.h"
+#include "taichi/runtime/metal/shaders/prolog.h"
 
 #ifdef TI_INSIDE_METAL_CODEGEN
 
@@ -73,4 +73,4 @@ METAL_END_AD_STACK_DEF
 #undef METAL_BEGIN_AD_STACK_DEF
 #undef METAL_END_AD_STACK_DEF
 
-#include "taichi/backends/metal/shaders/epilog.h"
+#include "taichi/runtime/metal/shaders/epilog.h"
