+++ taichi/codegen/metal/shaders/helpers.metal.h
@@ -1,4 +1,4 @@
-#include "taichi/backends/metal/shaders/prolog.h"
+#include "taichi/runtime/metal/shaders/prolog.h"
 
 #ifdef TI_INSIDE_METAL_CODEGEN
 
@@ -126,4 +126,4 @@ METAL_END_HELPERS_DEF
 #undef METAL_BEGIN_HELPERS_DEF
 #undef METAL_END_HELPERS_DEF
 
-#include "taichi/backends/metal/shaders/epilog.h"
+#include "taichi/runtime/metal/shaders/epilog.h"
