+++ taichi/codegen/metal/shaders/snode_bit_pointer.metal.h
@@ -1,4 +1,4 @@
-#include "taichi/backends/metal/shaders/prolog.h"
+#include "taichi/runtime/metal/shaders/prolog.h"
 
 #ifdef TI_INSIDE_METAL_CODEGEN
 
@@ -144,4 +144,4 @@ METAL_END_SRC_DEF
 #undef METAL_BEGIN_SRC_DEF
 #undef METAL_END_SRC_DEF
 
-#include "taichi/backends/metal/shaders/epilog.h"
+#include "taichi/runtime/metal/shaders/epilog.h"
