+++ taichi/codegen/metal/struct_metal.cpp
@@ -8,9 +8,9 @@
 #include <vector>
 
 #include "taichi/backends/metal/constants.h"
-#include "taichi/backends/metal/data_types.h"
-#include "taichi/backends/metal/features.h"
-#include "taichi/backends/metal/kernel_utils.h"
+#include "taichi/runtime/metal/data_types.h"
+#include "taichi/runtime/metal/features.h"
+#include "taichi/runtime/metal/kernel_utils.h"
 #include "taichi/math/arithmetic.h"
 #include "taichi/util/line_appender.h"
 
@@ -20,12 +20,12 @@ namespace metal {
 namespace {
 namespace shaders {
 #define TI_INSIDE_METAL_CODEGEN
-#include "taichi/backends/metal/shaders/runtime_structs.metal.h"
-#include "taichi/backends/metal/shaders/runtime_utils.metal.h"
-#include "taichi/backends/metal/shaders/snode_bit_pointer.metal.h"
+#include "taichi/runtime/metal/shaders/runtime_structs.metal.h"
+#include "taichi/runtime/metal/shaders/runtime_utils.metal.h"
+#include "taichi/codegen/metal/shaders/snode_bit_pointer.metal.h"
 #undef TI_INSIDE_METAL_CODEGEN
 
-#include "taichi/backends/metal/shaders/runtime_structs.metal.h"
+#include "taichi/runtime/metal/shaders/runtime_structs.metal.h"
 
 }  // namespace shaders
