+++ taichi/runtime/metal/aot_utils.h
@@ -2,7 +2,7 @@
 
 #include <vector>
 
-#include "taichi/backends/metal/kernel_utils.h"
+#include "taichi/runtime/metal/kernel_utils.h"
 
 namespace taichi {
 namespace lang {
