+++ taichi/runtime/metal/data_types.cpp
@@ -1,4 +1,4 @@
-#include "taichi/backends/metal/data_types.h"
+#include "taichi/runtime/metal/data_types.h"
 
 TLANG_NAMESPACE_BEGIN
 namespace metal {
