+++ taichi/runtime/metal/kernel_manager.h
@@ -6,8 +6,8 @@
 #include <vector>
 
 #include "taichi/backends/device.h"
-#include "taichi/backends/metal/kernel_utils.h"
-#include "taichi/backends/metal/struct_metal.h"
+#include "taichi/runtime/metal/kernel_utils.h"
+#include "taichi/codegen/metal/struct_metal.h"
 #include "taichi/lang_util.h"
 #include "taichi/program/compile_config.h"
 #include "taichi/program/kernel_profiler.h"
