+++ taichi/runtime/metal/shaders/print.metal.h
@@ -1,4 +1,4 @@
-#include "taichi/backends/metal/shaders/prolog.h"
+#include "taichi/runtime/metal/shaders/prolog.h"
 
 #ifdef TI_INSIDE_METAL_CODEGEN
 
@@ -210,4 +210,4 @@ METAL_END_PRINT_DEF
 #undef METAL_BEGIN_PRINT_DEF
 #undef METAL_END_PRINT_DEF
 
-#include "taichi/backends/metal/shaders/epilog.h"
+#include "taichi/runtime/metal/shaders/epilog.h"
