+++ taichi/runtime/program_impls/CMakeLists.txt
@@ -1,4 +1,5 @@
 # ./taichi/runtime/program_impls/CMakeLists.txt
 
-add_subdirectory(vulkan)
+add_subdirectory(metal)
 add_subdirectory(opengl)
+add_subdirectory(vulkan)
